working water

+++ src/renderer.cpp
@@ -52,11 +52,12 @@ int Renderer::IdxToCoord(int idx) {
     return idx * PIXEL_SIZE;
 }
 
+// these should really be size_t
 int Renderer::GetPixelAt(int* arr, int xi, int yi) {
     int x = IdxToCoord(xi);
     int y = IdxToCoord(yi);
 
-    size_t pos = x + y*SCREEN_WIDTH;
+    int pos = x + y*SCREEN_WIDTH;
     return arr[pos];
 }
